chore: Modularise `/admin` endpoints, improve Feedback endpoint docs (#2413)
