+++ api.planx.uk/modules/admin/docs.yaml
@@ -1,0 +1,28 @@
+openapi: 3.1.0
+info:
+  title: Plan✕ API
+  version: 0.1.0
+tags:
+  - name: admin
+    description: Admin only utility endpoints
+paths:
+  /admin/feedback:
+    get:
+      tags: ["admin"]
+      security:
+        - bearerAuth: []
+      summary: Download FeedbackFish CSV
+      parameters:
+        - in: query
+          name: cookie
+          type: string
+          description: Cookie from FeedbackFish to authenticate request
+          required: true
+      responses:
+        "200":
+          content:
+            text/csv:
+              schema:
+                type: string
+        "500":
+          $ref: "#/components/responses/ErrorMessage"
