+++ api.planx.uk/modules/admin/session/zip.test.ts
@@ -1,8 +1,8 @@
 import supertest from "supertest";
-import app from "../../server";
-import { authHeader } from "../../tests/mockJWT";
+import app from "../../../server";
+import { authHeader } from "../../../tests/mockJWT";
 
-jest.mock("../../send/exportZip", () => ({
+jest.mock("../../../send/exportZip", () => ({
   buildSubmissionExportZip: jest.fn().mockResolvedValue({
     filename: "tests/mocks/test.zip",
     remove: jest.fn,
